完成首页栏目分类查询出多少个月更新不及时以及空栏目的栏目

+++ src/main/java/com/trs/gov/kpi/entity/responsedata/UpdateNotInTimeChnl.java
@@ -7,6 +7,6 @@ import lombok.Data;
  */
 @Data
 public class UpdateNotInTimeChnl extends EmptyChnl{
-    private Integer countMonth;
+    private Double countMonth;
 
 }
